Increased efficiency of index sets

+++ GeometricAlgebraFulcrumLib/GeometricAlgebraFulcrumLib.Algebra/TensorAlgebra/Generic/Functions/DefaultWrappers.cs
@@ -110,7 +110,7 @@ namespace GeometricAlgebraFulcrumLib.Algebra.TensorAlgebra.Generic.Functions
         public T CreateOne()
         {
             if (typeof(T) == typeof(int))
-                return (T) (object) (int) 1;
+                return (T) (object) 1;
             else if (typeof(T) == typeof(float))
                 return (T) (object) (float) 1;
             if (typeof(T) == typeof(double))
@@ -124,7 +124,7 @@ namespace GeometricAlgebraFulcrumLib.Algebra.TensorAlgebra.Generic.Functions
         public T CreateZero()
         {
             if (typeof(T) == typeof(int))
-                return (T)(object)(int)0;
+                return (T)(object)0;
             else if (typeof(T) == typeof(float))
                 return (T)(object)(float)0;
             if (typeof(T) == typeof(double))
@@ -148,7 +148,7 @@ namespace GeometricAlgebraFulcrumLib.Algebra.TensorAlgebra.Generic.Functions
         public bool IsZero(T a)
         {
             if (typeof(T) == typeof(int))
-                return a.Equals((T)(object)(int)0);
+                return a.Equals((T)(object)0);
             else if (typeof(T) == typeof(float))
                 return a.Equals((T)(object)(float)0);
             if (typeof(T) == typeof(double))
